+++ Library/OcAudioLib/OcAudioLib.c
@@ -45,7 +45,7 @@ STATIC
 OC_AUDIO_PROTOCOL_PRIVATE
 mAudioProtocol = {
   .Signature       = OC_AUDIO_PROTOCOL_PRIVATE_SIGNATURE,
-  .AudioIo        = NULL,
+  .AudioIo         = NULL,
   .ProviderAcquire = NULL,
   .ProviderRelease = NULL,
   .ProviderContext = NULL,
@@ -58,6 +58,8 @@ mAudioProtocol = {
   .OcAudio         = {
     .Revision           = OC_AUDIO_PROTOCOL_REVISION,
     .Connect            = InternalOcAudioConnect,
+    .RawGainToDecibels  = InternalOcAudioRawGainToDecibels,
+    .SetDefaultGain     = InternalOcAudioSetDefaultGain,
     .SetProvider        = InternalOcAudioSetProvider,
     .PlayFile           = InternalOcAudioPlayFile,
     .StopPlayback       = InternalOcAudioStopPlayback,
@@ -158,75 +160,80 @@ OcAudioInstallProtocols (
   return &mAudioProtocol.OcAudio;
 }
 
-INT8
+VOID
 OcGetAmplifierGain (
-  OUT BOOLEAN  *Muted
+  OUT UINT8              *RawGain,
+  OUT INT8               *DecibelGain,
+  OUT BOOLEAN            *Muted,
+  OUT BOOLEAN            *TryConversion
   )
 {
-  EFI_STATUS  Status;
-  EFI_STATUS  AltStatus;
+  EFI_STATUS  Status1;
+  EFI_STATUS  Status2;
   UINTN       Size;
-  UINT8       Value;
-  INT8        DbValue;
-
-  *Muted = FALSE;
 
   //
-  // Check mute initially (only available in raw, amp-specific gain setting).
+  // Get mute setting and raw codec gain setting (all versions of macOS).
   //
-  Size   = sizeof (Value);
-  AltStatus = gRT->GetVariable (
+  Size    = sizeof (*RawGain);
+  Status1 = gRT->GetVariable (
     APPLE_SYSTEM_AUDIO_VOLUME_VARIABLE_NAME,
     &gAppleBootVariableGuid,
     NULL,
     &Size,
-    &Value
+    RawGain
     );
-  if (EFI_ERROR (AltStatus)) {
-    Value = 0;
-  } else if ((Value & APPLE_SYSTEM_AUDIO_VOLUME_MUTED) != 0) {
-    *Muted = TRUE;
-    Value &= APPLE_SYSTEM_AUDIO_VOLUME_VOLUME_MASK;
-  }
-
-  //
-  // If no mute, use dB gain setting (which can be applied to any amp on any codec).
-  //
-  if (*Muted) {
-    //
-    // SystemAudioVolumeDB appears not to contain sane values when SystemAudioVolume indicates muted,
-    // therefore we have to fall back to default gain value for use with audio assist, which never mutes.
-    //
-    DbValue = APPLE_SYSTEM_AUDIO_VOLUME_DB_MIN;
-    Status = EFI_NOT_STARTED;
+  if (!EFI_ERROR (Status1)) {
+    *Muted = (*RawGain & APPLE_SYSTEM_AUDIO_VOLUME_MUTED) != 0;
+    *RawGain &= APPLE_SYSTEM_AUDIO_VOLUME_VOLUME_MASK;
   } else {
     *Muted = FALSE;
-    Size   = sizeof (DbValue);
-    Status = gRT->GetVariable (
-      APPLE_SYSTEM_AUDIO_VOLUME_DB_VARIABLE_NAME,
-      &gAppleBootVariableGuid,
-      NULL,
-      &Size,
-      &DbValue
-      );
-    if (EFI_ERROR (Status)) {
-      DbValue = OC_AUDIO_DEFAULT_GAIN;
-    }
+    *RawGain = 0;
+  }
+
+  DEBUG ((
+    DEBUG_INFO,
+    "OCAU: System raw gain 0x%X, audio mute (for chime) %u - %r\n",
+    *RawGain,
+    *Muted,
+    Status1
+    ));
+
+  //
+  // Get dB gain setting, which can be correctly applied to any amp on any codec if available.
+  // (Not present at least in Lion 10.7 and earlier.)
+  //
+  Size    = sizeof (*DecibelGain);
+  Status2 = gRT->GetVariable (
+    APPLE_SYSTEM_AUDIO_VOLUME_DB_VARIABLE_NAME,
+    &gAppleBootVariableGuid,
+    NULL,
+    &Size,
+    DecibelGain
+    );
+  if (EFI_ERROR (Status2)) {
+    *DecibelGain = OC_AUDIO_DEFAULT_GAIN;
+  }
+
+  DEBUG ((
+    DEBUG_INFO,
+    "OCAU: System decibel gain %d dB%a - %r\n",
+    *DecibelGain,
+    *Muted ? " (probably invalid due to mute)" : "",
+    Status2
+    ));
+
+  //
+  // SystemAudioVolumeDB does not contain sane values when SystemAudioVolume indicates muted, so we have
+  // to fall back to default gain value or conversion; for use with audio assist, which never mutes.
+  //
+  if (*Muted) {
+    *DecibelGain   = OC_AUDIO_DEFAULT_GAIN;
+    Status2 = EFI_INVALID_PARAMETER;
   }
 
   //
-  // Log system saved raw gain for debugging purposes. Value derived later from dB gain should
-  // be the same to within about +/- 1, if applied on an amp with the same amp capabilities.
+  // If no saved decibel gain, but saved raw gain, it is worth trying to convert.
   //
-  DEBUG ((
-    DEBUG_INFO,
-    "OCAU: System amp gain %d dB, saved raw gain 0x%X, system audio mute (for chime) %u  - %r/%r\n",
-    DbValue,
-    Value,
-    *Muted,
-    AltStatus,
-    Status
-    ));
-
-  return DbValue;
+  *TryConversion = !EFI_ERROR (Status1) && EFI_ERROR (Status2);
 }